requirements.txt
google-cloud-texttospeech
python-dotenv